+++ lib/koala/http_service/request.rb
@@ -106,9 +106,7 @@ module Koala
       end
 
       def add_ssl_options(opts)
-        # require https if there's a token
-        return opts unless raw_args["access_token"]
-
+        # require https by default (can be overriden by explicitly setting other SSL options)
         {
           use_ssl: true,
           ssl: {verify: true}.merge(opts[:ssl] || {})